Restructure TLS Registry Extension
This commit reorganizes the TLS registry extension according to the guidelines discussed in https://github.com/quarkusio/quarkus/discussions/43130.

The restructuring addresses the issue logged in https://github.com/quarkusio/quarkus/issues/44711.

As a result of this change, extensions relying on the TLS registry build item will encounter a breaking change due to the updated package name.
