+++ extensions/tls-registry/deployment-spi/src/main/java/io/quarkus/tls/deployment/spi/TlsRegistryBuildItem.java
@@ -1,11 +1,14 @@
-package io.quarkus.tls;
+package io.quarkus.tls.deployment.spi;
 
 import java.util.function.Supplier;
 
 import io.quarkus.builder.item.SimpleBuildItem;
+import io.quarkus.tls.TlsConfigurationRegistry;
 
 /**
  * A build item that indicates that the TLS registry has been initialized.
+ * <p>
+ * Do not produce this build item elsewhere than in the TLS registry deployment processor.
  */
 public final class TlsRegistryBuildItem extends SimpleBuildItem {
