+++ extensions/tls-registry/deployment/src/test/java/io/quarkus/tls/TrustAllAndHostnameVerifierAlgorithmTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 import io.quarkus.test.QuarkusUnitTest;
-import io.quarkus.tls.runtime.keystores.TrustAllOptions;
+import io.quarkus.tls.runtime.TrustAllOptions;
 
 public class TrustAllAndHostnameVerifierAlgorithmTest {
