+++ extensions/tls-registry/pom.xml
@@ -16,6 +16,8 @@
     <packaging>pom</packaging>
     <modules>
         <module>deployment</module>
+        <module>deployment-spi</module>
+        <module>spi</module>
         <module>runtime</module>
         <module>cli</module>
     </modules>
