+++ extensions/tls-registry/runtime/src/main/java/io/quarkus/tls/runtime/VertxCertificateHolder.java
@@ -18,7 +18,6 @@ import javax.net.ssl.TrustManagerFactory;
 import io.quarkus.tls.TlsConfiguration;
 import io.quarkus.tls.runtime.config.TlsBucketConfig;
 import io.quarkus.tls.runtime.config.TlsConfigUtils;
-import io.quarkus.tls.runtime.keystores.TrustAllOptions;
 import io.vertx.core.Vertx;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.net.KeyCertOptions;
@@ -128,7 +127,7 @@ public class VertxCertificateHolder implements TlsConfiguration {
 
     @Override
     public boolean isTrustAll() {
-        return config().trustAll();
+        return config().trustAll() || getTrustStoreOptions() == TrustAllOptions.INSTANCE;
     }
 
     @Override
