+++ extensions/tls-registry/spi/src/main/java/io/quarkus/tls/BaseTlsConfiguration.java
@@ -98,4 +98,9 @@ public abstract class BaseTlsConfiguration implements TlsConfiguration {
     public boolean reload() {
         return false;
     }
+
+    @Override
+    public boolean isTrustAll() {
+        return false;
+    }
 }
