+++ extensions/tls-registry/spi/src/main/java/io/quarkus/tls/CertificateUpdatedEvent.java
